🚚 refactor(views): Update folder structure
