+++ src/views/DetailsForm/DetailsForm.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export default function DetailsForm({ initialData, onSubmit }) {
   return <div>DetailsForm</div>;
 }
